+++ client/Savitar2/src/worldDocument/WorldSettingsController.swift
@@ -11,6 +11,8 @@ import Cocoa
 class WorldSettingsController: NSViewController {
     var completionHandler: ((Bool, World?) -> Void)?
 
+    weak var sheetWindowController: WorldSettingsWindowController?
+
     var world: World? {
         get {
             return _editedWorld
@@ -18,14 +20,17 @@ class WorldSettingsController: NSViewController {
         set {
             // copy the world. We'll manipulate it in settings, and if the
             // user hits 'apply' we'll copy the changes back.
-            _editedWorld = newValue?.copy() as? World
+            guard let world = newValue?.copy() as? World else { return }
+            _editedWorld = world
 
             // we set the tab controllers' world here in the world setter
             // instead of prepare(for segue:) because prepare(for segue:)
             // gets called at storyboard instantiation, before we have
             // the chance to set the world value
-            for viewItem in (_tabViewController?.tabViewItems)! {
-                viewItem.viewController?.representedObject = _editedWorld!
+            if let tvc = _tabViewController {
+                for viewItem in tvc.tabViewItems {
+                    viewItem.viewController?.representedObject = _editedWorld!
+                }
             }
         }
     }
@@ -33,6 +38,9 @@ class WorldSettingsController: NSViewController {
     private var _editedWorld: World?
     private var _tabViewController: NSTabViewController?
     private var tabIndexObservation: NSKeyValueObservation?
+    private var didPolishChrome = false
+
+    private let footerHeight: CGFloat = 72
 
     override func prepare(for segue: NSStoryboardSegue, sender _: Any?) {
         // grab a reference to the tabViewController, we'll use it in the
@@ -44,12 +52,19 @@ class WorldSettingsController: NSViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
+        sheetWindowController = view.window?.windowController as? WorldSettingsWindowController
+        sheetWindowController?.settingsController = self
         if let tvc = _tabViewController {
             tabIndexObservation = tvc.observe(\.selectedTabViewItemIndex) { [weak self] _, _ in
-                self?.updateContextualHelpForSelectedTab()
+                self?.tabSelectionDidChange()
             }
         }
-        updateContextualHelpForSelectedTab()
+        tabSelectionDidChange()
+    }
+
+    override func viewWillAppear() {
+        super.viewWillAppear()
+        polishChromeForHIGIfNeeded()
     }
 
     deinit {
@@ -64,6 +79,53 @@ class WorldSettingsController: NSViewController {
         completionHandler?(false, nil)
     }
 
+    func fittingContentSize(for tab: SavitarHelp.WorldSettingsTab) -> NSSize {
+        let width: CGFloat = 480
+        let preferredTabHeight = tab.preferredSheetHeight
+        var tabHeight = preferredTabHeight
+        if let tabView = _tabViewController?.tabViewItems[tab.rawValue].view {
+            tabView.layoutSubtreeIfNeeded()
+            let fitted = tabView.fittingSize
+            if fitted.height > 1 {
+                tabHeight = max(fitted.height, preferredTabHeight)
+            }
+        }
+        return NSSize(width: max(width, tabViewWidth(for: tab)), height: tabHeight + footerHeight)
+    }
+
+    private func tabViewWidth(for tab: SavitarHelp.WorldSettingsTab) -> CGFloat {
+        guard let tabView = _tabViewController?.tabViewItems[tab.rawValue].view else { return 0 }
+        return tabView.fittingSize.width
+    }
+
+    private func tabSelectionDidChange() {
+        polishChromeForHIGIfNeeded()
+        updateContextualHelpForSelectedTab()
+        let tab = selectedWorldSettingsTab ?? .starting
+        sheetWindowController?.updateForTab(tab, animated: view.window?.isVisible == true)
+    }
+
+    private func polishChromeForHIGIfNeeded() {
+        guard !didPolishChrome else { return }
+        didPolishChrome = true
+
+        for subview in view.subviews {
+            if let field = subview as? NSTextField,
+               field.cell?.title == "Settings",
+               field.alignment == .center {
+                field.isHidden = true
+                for constraint in view.constraints where
+                    (constraint.firstItem as? NSObject) === field && constraint.firstAttribute == .height {
+                    constraint.constant = 0
+                }
+            }
+            if let button = subview as? NSButton, button.title == "Apply" {
+                button.title = "OK"
+                button.keyEquivalent = "\r"
+            }
+        }
+    }
+
     private func updateContextualHelpForSelectedTab() {
         let tab = selectedWorldSettingsTab ?? .starting
         SavitarHelpButton.installInTopTrailingCorner(of: view, for: .worldSettings(tab))
@@ -72,5 +134,21 @@ class WorldSettingsController: NSViewController {
     private var selectedWorldSettingsTab: SavitarHelp.WorldSettingsTab? {
         guard let index = _tabViewController?.selectedTabViewItemIndex else { return nil }
         return SavitarHelp.WorldSettingsTab(rawValue: index)
+    }
+
+    var currentWorldSettingsTab: SavitarHelp.WorldSettingsTab {
+        selectedWorldSettingsTab ?? .starting
+    }
+}
+
+private extension SavitarHelp.WorldSettingsTab {
+    var preferredSheetHeight: CGFloat {
+        switch self {
+        case .starting: return 400
+        case .appearance: return 480
+        case .input: return 480
+        case .output: return 300
+        case .closing: return 380
+        }
     }
 }