+++ client/Savitar2Tests/WorldTests.swift
@@ -224,6 +224,29 @@ class WorldTests: XCTestCase {
         XCTAssertTrue(xmlOutString.contains("FLAGS=\"ansi+html+CROnly\""))
     }
 
+    func testAutoCloseFlagRoundTripThroughXML() throws {
+        let xmlString = """
+        <WORLD
+            NAME="Alter Aeon"
+            URL="telnet://example.com:3000"
+            FLAGS="html+ansi+autoClose"
+            CMDMARKER="##"
+            VARMARKER="%%"
+            WILDMARKER="$$"
+            RESOLUTION="80x24x2"
+        />
+        """
+
+        let xml = try XML.parse(xmlString)
+        let world = World()
+        try world.parse(xml: xml[WorldElemIdentifier])
+
+        XCTAssertTrue(world.flags.contains(.autoClose))
+
+        let xmlOutString = try world.toXMLElement().xmlString
+        XCTAssertTrue(xmlOutString.contains("autoClose"))
+    }
+
     func testCommandLinePostfixDefaultsToCRLF() {
         let world = World()
         XCTAssertEqual(world.commandLinePostfix, "\r\n")
@@ -387,6 +410,23 @@ class StickyCommandInputTests: XCTestCase {
         XCTAssertEqual(inputController.commandHistory(), ["look"])
         XCTAssertEqual(inputController.cmdBuf.count, 2)
         XCTAssertEqual(inputController.cmdIndex, 2)
+    }
+}
+
+class SessionLogoffTests: XCTestCase {
+    func testLogoffLinesToSendSplitsMultipleLines() {
+        let world = World()
+        world.logoffCmd = "quit\n@quit\n"
+        let session = Session(world: world, sessionHandler: MockSessionHandler())
+
+        XCTAssertEqual(session.logoffLinesToSend(), ["quit", "@quit"])
+    }
+
+    func testLogoffLinesToSendEmptyWhenUnset() {
+        let world = World()
+        let session = Session(world: world, sessionHandler: MockSessionHandler())
+
+        XCTAssertEqual(session.logoffLinesToSend(), [])
     }
 }